Changes on the event mechanism
Basically added the event listener function to NOT have to add a messageType attribute to each DOM element, child of the element responsible for this specific type of messages. For example, since each repository element is responsible for the messages it receives, each children do not need a messageType attribute. I traverse the DOM, to find the element responsible for messaging without using e.composedPath() because it is not universally implemented browser-wise.